+++ DSCResources/MSFT_xExchMaintenanceMode/MSFT_xExchMaintenanceMode.psm1
@@ -74,10 +74,10 @@ function Get-TargetResource
 
     Write-FunctionEntry -Parameters @{'Enabled' = $Enabled} -Verbose:$VerbosePreference
 
-    #Load TransportMaintenanceMode Helper
+    # Load TransportMaintenanceMode Helper
     Import-Module "$((Get-Item -LiteralPath "$($PSScriptRoot)"))\TransportMaintenance.psm1" -Verbose:0
 
-    #Establish remote Powershell session
+    # Establish remote Powershell session
     Get-RemoteExchangeSession -Credential $Credential -CommandsToLoad 'Get-*' -Verbose:$VerbosePreference
 
     $maintenanceModeStatus = GetMaintenanceModeStatus -EnteringMaintenanceMode $Enabled -DomainController $DomainController
@@ -85,10 +85,10 @@ function Get-TargetResource
 
     if ($null -ne $maintenanceModeStatus)
     {
-        #Determine which components are Active
-        $activeComponents = $MaintenanceModeStatus.ServerComponentState | Where-Object {$_.State -eq "Active"}
+        # Determine which components are Active
+        $activeComponents = $maintenanceModeStatus.ServerComponentState | Where-Object -FilterScript {$_.State -eq "Active"}
 
-        [System.String[]]$activeComponentsList = @()
+        [System.String[]] $activeComponentsList = @()
 
         if ($null -ne $activeComponents)
         {
@@ -100,15 +100,13 @@ function Get-TargetResource
 
         $activeComponentCount = $activeComponentsList.Count
 
-
-        #Figure out what our Enabled state should really be in case UpgradedServerVersion was passed
+        # Figure out what our Enabled state should really be in case UpgradedServerVersion was passed
         $isEnabled = $Enabled
 
         if ($Enabled -eq $true -and $atDesiredVersion -eq $true)
         {
             $isEnabled = $false
         }
-
 
         $returnValue = @{
             Enabled              = [System.Boolean] $isEnabled
@@ -237,8 +235,8 @@ function Set-TargetResource
     if ($null -ne $maintenanceModeStatus)
     {
         #Set vars relevant to both 'Enabled' code paths
-        $htStatus = $MaintenanceModeStatus.ServerComponentState | Where-Object {$_.Component -eq "HubTransport"}
-        $haStatus = $MaintenanceModeStatus.ServerComponentState | Where-Object {$_.Component -eq "HubTransport"}
+        $htStatus = $MaintenanceModeStatus.ServerComponentState | Where-Object -FilterScript {$_.Component -eq "HubTransport"}
+        $haStatus = $MaintenanceModeStatus.ServerComponentState | Where-Object -FilterScript {$_.Component -eq "HubTransport"}
 
         #Put the server into maintenance mode
         if ($Enabled -eq $true)
@@ -370,7 +368,7 @@ function Set-TargetResource
                     if ((IsComponentCheckedByDefault -ComponentName $component) -eq $false)
                     {
                         $status = $null
-                        $status = $MaintenanceModeStatus.ServerComponentState | Where-Object {$_.Component -like "$($component)"}
+                        $status = $MaintenanceModeStatus.ServerComponentState | Where-Object -FilterScript {$_.Component -like "$($component)"}
 
                         if ($null -ne $status -and $status.State -ne 'Active')
                         {
@@ -515,7 +513,7 @@ function Test-TargetResource
                     $testResults = $false
                 }
 
-                if ($null -ne ($MaintenanceModeStatus.ServerComponentState | Where-Object {$_.State -ne "Inactive" -and $_.Component -ne "Monitoring" -and $_.Component -ne "RecoveryActionsEnabled"}))
+                if ($null -ne ($MaintenanceModeStatus.ServerComponentState | Where-Object -FilterScript {$_.State -ne "Inactive" -and $_.Component -ne "Monitoring" -and $_.Component -ne "RecoveryActionsEnabled"}))
                 {
                     Write-Verbose -Message 'One or more components have a status other than Inactive'
                     $testResults = $false
@@ -564,7 +562,7 @@ function Test-TargetResource
         #Make sure the server is fully out of maintenance mode
         else
         {
-            $activeComponents = $MaintenanceModeStatus.ServerComponentState | Where-Object {$_.State -eq "Active"}
+            $activeComponents = $MaintenanceModeStatus.ServerComponentState | Where-Object -FilterScript {$_.State -eq "Active"}
 
             if ($null -eq $activeComponents)
             {
@@ -572,7 +570,7 @@ function Test-TargetResource
                 $testResults = $false
             }
 
-            if ($null -eq ($activeComponents | Where-Object {$_.Component -eq "ServerWideOffline"}))
+            if ($null -eq ($activeComponents | Where-Object -FilterScript {$_.Component -eq "ServerWideOffline"}))
             {
                 Write-Verbose -Message 'Component ServerWideOffline is not Active'
                 $testResults = $false
@@ -580,14 +578,14 @@ function Test-TargetResource
 
             if ($serverVersion -in '2013', '2016')
             {
-                if ($null -eq ($activeComponents | Where-Object {$_.Component -eq "UMCallRouter"}))
+                if ($null -eq ($activeComponents | Where-Object -FilterScript {$_.Component -eq "UMCallRouter"}))
                 {
                     Write-Verbose -Message 'Component UMCallRouter is not Active'
                     $testResults = $false
                 }
             }
 
-            if ($null -eq ($activeComponents | Where-Object {$_.Component -eq "HubTransport"}))
+            if ($null -eq ($activeComponents | Where-Object -FilterScript {$_.Component -eq "HubTransport"}))
             {
                 Write-Verbose -Message 'Component HubTransport is not Active'
                 $testResults = $false
@@ -605,13 +603,13 @@ function Test-TargetResource
                 $testResults = $false
             }
 
-            if ($null -eq ($activeComponents | Where-Object {$_.Component -eq "Monitoring"}))
+            if ($null -eq ($activeComponents | Where-Object -FilterScript {$_.Component -eq "Monitoring"}))
             {
                 Write-Verbose -Message 'Component Monitoring is not Active'
                 $testResults = $false
             }
 
-            if ($null -eq ($activeComponents | Where-Object {$_.Component -eq "RecoveryActionsEnabled"}))
+            if ($null -eq ($activeComponents | Where-Object -FilterScript {$_.Component -eq "RecoveryActionsEnabled"}))
             {
                 Write-Verbose -Message 'Component RecoveryActionsEnabled is not Active'
                 $testResults = $false
@@ -624,7 +622,7 @@ function Test-TargetResource
                     if ((IsComponentCheckedByDefault -ComponentName $component) -eq $false)
                     {
                         $status = $null
-                        $status = $MaintenanceModeStatus.ServerComponentState | Where-Object {$_.Component -like "$($component)"}
+                        $status = $MaintenanceModeStatus.ServerComponentState | Where-Object -FilterScript {$_.Component -like "$($component)"}
 
                         if ($null -ne $status -and $Status.State -ne "Active")
                         {
@@ -707,7 +705,7 @@ function GetQueueMessageCount
 
     if ($null -ne $MaintenanceModeStatus.Queues)
     {
-        foreach ($queue in $MaintenanceModeStatus.Queues | Where-Object {$_.Identity -notlike "*\Shadow\*"})
+        foreach ($queue in $MaintenanceModeStatus.Queues | Where-Object -FilterScript {$_.Identity -notlike "*\Shadow\*"})
         {
             Write-Verbose -Message "Found queue '$($queue.Identity)' with a message count of '$($queue.MessageCount)'."
             $messageCount += $queue.MessageCount
@@ -740,7 +738,7 @@ function GetActiveDBCount
     [UInt32]$activeDBCount = 0
 
     #Get DB's with a status of Mounted, Mounting, Dismounted, or Dismounting
-    $localDBs = $MaintenanceModeStatus.DBCopyStatus | Where-Object {$_.Status -like "Mount*" -or $_.Status -like "Dismount*"}
+    $localDBs = $MaintenanceModeStatus.DBCopyStatus | Where-Object -FilterScript {$_.Status -like "Mount*" -or $_.Status -like "Dismount*"}
 
     #Ensure that any DB's we found actually have copies
     foreach ($db in $localDBs)
@@ -1066,7 +1064,7 @@ function ChangeComponentState
 
     [System.Boolean]$madeChange = $false
 
-    $componentState = $MaintenanceModeStatus.ServerComponentState | Where-Object {$_.Component -like "$($Component)"}
+    $componentState = $MaintenanceModeStatus.ServerComponentState | Where-Object -FilterScript {$_.Component -like "$($Component)"}
 
     if ($null -ne $componentState)
     {
@@ -1086,7 +1084,7 @@ function ChangeComponentState
             if ($State -eq "Active" -and $SetInactiveComponentsFromAnyRequesterToActive -eq $true)
             {
                 $additionalRequesters = $null
-                $additionalRequesters = $componentState.LocalStates | Where-Object {$_.Requester -notlike "$($Requester)" -and $_.State -notlike "Active"}
+                $additionalRequesters = $componentState.LocalStates | Where-Object -FilterScript {$_.Requester -notlike "$($Requester)" -and $_.State -notlike "Active"}
 
                 if ($null -ne $additionalRequesters)
                 {